[MOVE] 맵페이지 폴더 생성
